Move the KSP test to astronomy.
