+++ astronomy/astronomy.vcxproj
@@ -224,6 +224,8 @@
   <ItemGroup>
     <ClCompile Include="..\base\status.cpp" />
     <ClCompile Include="date_time_test.cpp" />
+    <ClCompile Include="ksp_resonance_test.cpp" />
+    <ClCompile Include="ksp_system_test.cpp" />
     <ClCompile Include="solar_system_dynamics_test.cpp" />
     <ClCompile Include="time_scales_test.cpp" />
     <ClCompile Include="lunar_eclipse_test.cpp" />
