+++ astronomy/astronomy.vcxproj.filters
@@ -143,5 +143,11 @@
     <ClCompile Include="молния_orbit_test.cpp">
       <Filter>Test Files</Filter>
     </ClCompile>
+    <ClCompile Include="ksp_resonance_test.cpp">
+      <Filter>Test Files</Filter>
+    </ClCompile>
+    <ClCompile Include="ksp_system_test.cpp">
+      <Filter>Test Files</Filter>
+    </ClCompile>
   </ItemGroup>
 </Project>
